Next State, Next cell and Next Repetition Solved
This patch solve the interruption problem that didnt let the nextState, nextRepetition and nextCell work the way it was intended.

Future patches will be focus on cleanin the code, properly commented and quality of life updates such as the implementation of default values for bothe NMH and LI k values for the pid.